fixed 64 bit compatibility problems
git-svn-id: https://svn.urbansim.org/trunk@11470 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ opus_core/choice_model.py
@@ -198,7 +198,7 @@ class ChoiceModel(ChunkModel):
                             index[x,coef[submodel].get_equations_index()[res[x]]], res_positive_idx))
                 else:
                     restmp[res_positive_idx] = coef[submodel].get_equations_index()[res[res_positive_idx]]
-                result[self.observations_mapping[submodel]] = restmp
+                result[self.observations_mapping[submodel]] = restmp.astype(int32)
         return result
 
     def simulate_submodel(self, data, coefficients, submodel=0):
